requirements.txt
flask
flask-sqlalchemy
flask-cors
pymysql
python-dotenv
flask-jwt-extended
Flask-Mail
